+++ CHANGELOG.md
@@ -127,6 +127,7 @@
 - Manage origin date on supplier invoice. This date is used for compute due date and is used in account move line to fill the origin date.
 - Fill analytic template on move line when we generate the accounting move from an invoice and use it as a key to consolidate the generated move lines.
 - Fill analytic template on invoice line when we generate the invoice from a sale or purchase order or stock move linked to a sale or purchase order.
+- STOCK MOVE : filter on fromStockLocation
 
 ## Improvements
 - Move project resource menu entry to project configuration menu.